# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/loop-presets/{id}": {
      "get": {
        "operationId": "LoopPresetsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "LoopPresets"
        ],
        "x-required-role": "QA_MANAGER"
      },
      "delete": {
        "operationId": "LoopPresetsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "LoopPresets"
        ],
        "x-required-role": "QA_MANAGER"
      }
    },
...
    "/capture-points/{id}": {
      "delete": {
        "operationId": "CapturePointsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CapturePoints"
        ],
        "x-required-role": "QA_MANAGER"
      }
    },
...
  },
  "info": {
    "title": "Inspect API",
    "description": "Tamper-proof, AQL-driven pre-shipment QC inspection platform. Every route is guarded by default; `x-required-role` on each operation records the additive-hierarchy floor (INSPECTOR < QA_MANAGER < ORG_OWNER < PLATFORM_ADMIN), or \"public\" for unauthenticated routes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}